Elias/Test_Share_AML-Endpoint-Testing-Tool/requirements.txt
Flask==3.0.3
pillow==11.1.0
flask-swagger-ui==4.11.1